# docs/openapi.yaml
openapi: 3.1.0
info:
  title: PayCat API
  description: |
    Unified Payment & Subscription Management System.
    A RevenueCat-compatible API for managing in-app purchases across iOS, Android, and Web (Stripe).
  version: 1.0.0
  contact:
    name: PayCat Support
  license:
    name: MIT
servers:
  - url: https://your-worker.workers.dev
    description: Production
  - url: http://localhost:8787
    description: Local Development
...
security:
  - ApiKeyAuth: []
paths:
...
  /v1/offerings:
    get:
      tags: [Offerings]
      summary: Get offerings
      description: Get current offerings configuration for the requesting user
      parameters:
        - name: app_user_id
          in: query
          schema:
            type: string
          description: User ID for targeted offerings
      responses:
        '200':
          description: Offerings configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OfferingsResponse'

    post:
      tags: [Offerings]
      summary: Create offering
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [identifier]
              properties:
                identifier:
                  type: string
                display_name:
                  type: string
                description:
                  type: string
                is_current:
                  type: boolean
                metadata:
                  type: object
      responses:
        '201':
          description: Offering created
  /v1/offerings/{identifier}:
    get:
      tags: [Offerings]
      summary: Get offering by identifier
      parameters:
        - name: identifier
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Offering details
  /v1/products:
    get:
      tags: [Offerings]
      summary: List products
      responses:
        '200':
          description: List of products

    post:
      tags: [Offerings]
      summary: Create product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductCreate'
      responses:
        '201':
          description: Product created
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for your app (pk_live_xxx or pk_test_xxx)
    BearerAuth:
      type: http
      scheme: bearer
      description: Admin session token
  schemas:
...
    OfferingsResponse:
      type: object
      properties:
        current_offering_id:
          type: string
        offerings:
          type: array
          items:
            $ref: '#/components/schemas/Offering'
    Offering:
      type: object
      properties:
        identifier:
          type: string
        description:
          type: string
        packages:
          type: array
          items:
            $ref: '#/components/schemas/Package'
    Package:
      type: object
      properties:
        identifier:
          type: string
        package_type:
          type: string
          enum: [monthly, annual, weekly, lifetime, custom]
        product:
          $ref: '#/components/schemas/Product'
    Product:
      type: object
      properties:
        store_product_id:
          type: string
        platform:
          type: string
        product_type:
          type: string
        price:
          type: object
          properties:
            amount:
              type: number
            currency:
              type: string
    ProductCreate:
      type: object
      required: [store_product_id, platform, product_type]
      properties:
        store_product_id:
          type: string
        platform:
          type: string
          enum: [ios, android, stripe]
        product_type:
          type: string
          enum: [subscription, consumable, non_consumable]
        display_name:
          type: string
        description:
          type: string
        default_price_amount:
          type: integer
        default_price_currency:
          type: string
        subscription_period:
          type: string
        trial_period:
          type: string
...